+++ celery/contrib/pytest.py
@@ -78,6 +78,7 @@ def celery_session_app(request,
 def celery_session_worker(request,
                           celery_session_app,
                           celery_includes,
+                          celery_class_tasks,
                           celery_worker_pool,
                           celery_worker_parameters):
     # type: (Any, Celery, Sequence[str], str, Any) -> WorkController
@@ -85,6 +86,8 @@ def celery_session_worker(request,
     if not NO_WORKER:
         for module in celery_includes:
             celery_session_app.loader.import_task_module(module)
+        for class_task in celery_class_tasks:
+            celery_session_app.tasks.register(class_task)
         with worker.start_worker(celery_session_app,
                                  pool=celery_worker_pool,
                                  **celery_worker_parameters) as w:
@@ -171,7 +174,7 @@ def celery_app(request,
         yield app
 
 
-@pytest.fixture()
+@pytest.fixture(scope='session')
 def celery_class_tasks():
     return []
 
@@ -180,7 +183,6 @@ def celery_class_tasks():
 def celery_worker(request,
                   celery_app,
                   celery_includes,
-                  celery_class_tasks,
                   celery_worker_pool,
                   celery_worker_parameters):
     # type: (Any, Celery, Sequence[str], str, Any) -> WorkController
@@ -188,8 +190,6 @@ def celery_worker(request,
     if not NO_WORKER:
         for module in celery_includes:
             celery_app.loader.import_task_module(module)
-        for class_task in celery_class_tasks:
-            celery_app.tasks.register(class_task)
         with worker.start_worker(celery_app,
                                  pool=celery_worker_pool,
                                  **celery_worker_parameters) as w:
