+++ t/integration/test_tasks.py
@@ -2,36 +2,20 @@ from __future__ import absolute_import, unicode_literals
 
 import pytest
 
-from celery import group, Task
+from celery import group
 
 from .conftest import get_active_redis_channels
 from .tasks import (add, add_ignore_result, print_unicode, retry_once,
-                    retry_once_priority, sleeping)
-
-
-class ClassBasedAutoRetryTask(Task):
-    name = 'auto_retry_class_task'
-    autoretry_for = (ValueError,)
-    retry_kwargs = {'max_retries': 1}
-    retry_backoff = True
-
-    def run(self):
-        if self.request.retries:
-            return self.request.retries
-        raise ValueError()
-
-
-@pytest.fixture()
-def celery_class_tasks():
-    return [ClassBasedAutoRetryTask]
+                    retry_once_priority, sleeping, ClassBasedAutoRetryTask)
 
 
 class test_class_based_tasks:
 
     @pytest.mark.flaky(reruns=5, reruns_delay=2)
-    def test_class_based_task_retried(self, celery_app, celery_worker):
+    def test_class_based_task_retried(self, celery_session_app,
+                                      celery_session_worker):
         task = ClassBasedAutoRetryTask()
-        celery_app.tasks.register(task)
+        celery_session_app.tasks.register(task)
         res = task.delay()
         assert res.get(timeout=10) == 1
